Запустил, но ничего не работает.

+++ Assets/Genome.cs
@@ -7,22 +7,22 @@ public class Genome
     private Random random = new Random();
     public static int skillCount = 5;
 
-    public double[] weights;
+    public float[] weights;
     public int[] skills;
 
     public Genome(int size)
     {
-        weights = new double[size];
+        weights = new float[size];
         skills = new int[skillCount];
         for (int i = 0; i < size; i++)
         {
-            weights[i] = random.NextDouble();
+            weights[i] = (float)random.NextDouble();
         }
     }
 
     public Genome(Genome a)
     {
-        weights = new double[a.weights.Length];
+        weights = new float[a.weights.Length];
         Array.Copy(a.weights, 0, weights, 0, a.weights.Length);
         skills = new int[skillCount];
         Array.Copy(a.skills, 0, skills, 0, skillCount);
@@ -32,7 +32,7 @@ public class Genome
     {
         for (int i = 0; i < weights.Length; i++)
         {
-            if (random.NextDouble() < 0.1) weights[i] += random.NextDouble();
+            if (random.NextDouble() < 0.1) weights[i] += (float)random.NextDouble();
         }
         for (int i = 0; i < skillCount; i++)
         {
