+++ Assets/Layer.cs
@@ -4,13 +4,13 @@ using System.Collections.Generic;
 public class Layer
 {
     public int size;
-    public double[] neurons;
-    public double[,] weights;
+    public float[] neurons;
+    public float[,] weights;
 
     public Layer(int size, int nextSize)
     {
         this.size = size;
-        neurons = new double[size];
-        weights = new double[size, nextSize];
+        neurons = new float[size];
+        weights = new float[size, nextSize];
     }
 }
